+++ test/libyul/YulOptimizerTestCommon.cpp
@@ -311,6 +311,13 @@ YulOptimizerTestCommon::YulOptimizerTestCommon(std::shared_ptr<Object const> _ob
 			ExpressionJoiner::run(*m_context, block);
 			return block;
 		}},
+		{"unusedStoreEliminatorNoSsaTransform", [&]() {
+			auto block = disambiguate();
+			updateContext(block);
+			ForLoopInitRewriter::run(*m_context, block);
+			UnusedStoreEliminator::run(*m_context, block);
+			return block;
+		}},
 		{"equalStoreEliminator", [&]() {
 			auto block = disambiguate();
 			updateContext(block);